Users_Backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.6.0
redis
requests==2.32.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
typing-extensions>=4.7.1